refactor: split large Android files by responsibility
