+++ app/src/main/java/com/vangeaux/lagrange/BookOrbitApp.kt
@@ -443,281 +443,6 @@ private fun StartupScreen(message: String) {
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-private fun ServerSetupScreen(
-    initialServerUrl: String,
-    message: String?,
-    onContinue: (String) -> Unit
-) {
-    var server by remember(initialServerUrl) { mutableStateOf(initialServerUrl) }
-    var error by remember { mutableStateOf<String?>(null) }
-
-    Scaffold(
-        topBar = { BookOrbitTopBar(title = "Connect") },
-        containerColor = MaterialTheme.colorScheme.background
-    ) { padding ->
-        Box(
-            modifier = Modifier
-                .padding(padding)
-                .fillMaxSize(),
-            contentAlignment = Alignment.TopCenter
-        ) {
-            Column(
-                modifier = Modifier
-                    .widthIn(max = 560.dp)
-                    .fillMaxWidth()
-                    .padding(horizontal = 24.dp, vertical = 32.dp),
-                verticalArrangement = Arrangement.spacedBy(16.dp)
-            ) {
-                OrbitEyebrow("Private reader")
-                Text("Your library, in orbit.", style = MaterialTheme.typography.displaySmall)
-                Text(
-                    "Connect securely to your BookOrbit server. Your library stays on your server; this app is your reading window.",
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    style = MaterialTheme.typography.bodyLarge
-                )
-                if (!message.isNullOrBlank()) {
-                    OrbitMessage(message, tone = OrbitMessageTone.ERROR)
-                }
-                OutlinedTextField(
-                    value = server,
-                    onValueChange = {
-                        server = it
-                        error = null
-                    },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .semantics { contentDescription = "BookOrbit server URL" },
-                    label = { Text("Server URL") },
-                    placeholder = { Text("https://books.example.com") },
-                    singleLine = true
-                )
-                error?.let {
-                    OrbitMessage(it, tone = OrbitMessageTone.ERROR)
-                }
-                Button(
-                    onClick = {
-                        val normalized = normalizeServerUrl(server)
-                        if (normalized == null) {
-                            error = invalidServerUrlMessage()
-                        } else {
-                            onContinue(normalized)
-                        }
-                    },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .heightIn(min = 52.dp)
-                ) {
-                    Text("Continue")
-                }
-                if (server.isNotBlank() && !message.isNullOrBlank()) {
-                    OutlinedButton(
-                        onClick = { onContinue(server) },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .heightIn(min = 52.dp)
-                    ) {
-                        Text("Retry")
-                    }
-                }
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun LoginScreen(
-    serverUrl: String,
-    message: String?,
-    isSubmitting: Boolean,
-    onChangeServer: () -> Unit,
-    onSubmit: (String, String) -> Unit,
-    onOpenServerSignIn: () -> Unit
-) {
-    var username by remember(serverUrl) { mutableStateOf("") }
-    var password by remember(serverUrl) { mutableStateOf("") }
-    var passwordVisible by remember(serverUrl) { mutableStateOf(false) }
-    var validationMessage by remember(serverUrl) { mutableStateOf<String?>(null) }
-    var showServerSignInInfo by remember(serverUrl) { mutableStateOf(false) }
-    val submit = {
-        when {
-            username.isBlank() -> validationMessage = "Enter your username."
-            password.isBlank() -> validationMessage = "Enter your password."
-            else -> {
-                validationMessage = null
-                onSubmit(username.trim(), password)
-            }
-        }
-    }
-    Scaffold(
-        topBar = {
-            BookOrbitTopBar(
-                title = "Sign in",
-                actions = { TextButton(onClick = onChangeServer) { Text("Change server") } }
-            )
-        },
-        containerColor = MaterialTheme.colorScheme.background
-    ) { padding ->
-        Column(
-            modifier = Modifier
-                .padding(padding)
-                .padding(horizontal = 12.dp, vertical = 12.dp)
-                .fillMaxSize()
-                .verticalScroll(rememberScrollState())
-        ) {
-            if (!message.isNullOrBlank()) {
-                OrbitMessage(
-                    text = message,
-                    modifier = Modifier.padding(bottom = 12.dp),
-                    tone = OrbitMessageTone.ERROR
-                )
-            }
-            Surface(
-                modifier = Modifier.fillMaxWidth(),
-                shape = MaterialTheme.shapes.large,
-                tonalElevation = 1.dp
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(20.dp),
-                    verticalArrangement = Arrangement.spacedBy(14.dp)
-                ) {
-                    Column(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .semantics(mergeDescendants = true) {},
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        Text(
-                            text = "Lagrange",
-                            style = MaterialTheme.typography.headlineSmall,
-                            modifier = Modifier.semantics { heading() }
-                        )
-                        Text(
-                            text = "a BookOrbit reader",
-                            style = MaterialTheme.typography.bodySmall,
-                            color = MaterialTheme.colorScheme.secondary
-                        )
-                    }
-                    OrbitEyebrow("BookOrbit server")
-                    Text(serverUrl, style = MaterialTheme.typography.bodySmall)
-                    OutlinedTextField(
-                        value = username,
-                        onValueChange = {
-                            username = it
-                            validationMessage = null
-                        },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .semantics { contentDescription = "BookOrbit username" },
-                        label = { Text("Username") },
-                        singleLine = true,
-                        enabled = !isSubmitting,
-                        keyboardOptions = androidx.compose.foundation.text.KeyboardOptions(
-                            keyboardType = KeyboardType.Text,
-                            imeAction = ImeAction.Next
-                        )
-                    )
-                    OutlinedTextField(
-                        value = password,
-                        onValueChange = {
-                            password = it
-                            validationMessage = null
-                        },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .semantics { contentDescription = "BookOrbit password" },
-                        label = { Text("Password") },
-                        singleLine = true,
-                        enabled = !isSubmitting,
-                        visualTransformation = if (passwordVisible) {
-                            VisualTransformation.None
-                        } else {
-                            PasswordVisualTransformation()
-                        },
-                        trailingIcon = {
-                            TextButton(
-                                onClick = { passwordVisible = !passwordVisible },
-                                enabled = !isSubmitting
-                            ) {
-                                Text(if (passwordVisible) "Hide" else "Show")
-                            }
-                        },
-                        keyboardOptions = androidx.compose.foundation.text.KeyboardOptions(
-                            keyboardType = KeyboardType.Password,
-                            imeAction = ImeAction.Done
-                        ),
-                        keyboardActions = androidx.compose.foundation.text.KeyboardActions(onDone = { submit() })
-                    )
-                    validationMessage?.let {
-                        OrbitMessage(it, tone = OrbitMessageTone.ERROR)
-                    }
-                    Button(
-                        onClick = submit,
-                        enabled = !isSubmitting,
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .heightIn(min = 52.dp)
-                    ) {
-                        if (isSubmitting) {
-                            CircularProgressIndicator(
-                                modifier = Modifier.size(20.dp),
-                                strokeWidth = 2.dp
-                            )
-                        } else {
-                            Text("Sign in")
-                        }
-                    }
-                    HorizontalDivider(modifier = Modifier.padding(vertical = 4.dp))
-                    Text(
-                        text = "BookOrbit controls which server sign-in options are available.",
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.outline
-                    )
-                    Row(
-                        modifier = Modifier.fillMaxWidth(),
-                        verticalAlignment = Alignment.CenterVertically
-                    ) {
-                        OutlinedButton(
-                            onClick = onOpenServerSignIn,
-                            enabled = !isSubmitting,
-                            modifier = Modifier
-                                .weight(1f)
-                                .heightIn(min = 52.dp)
-                        ) {
-                            Text("Open server sign-in")
-                        }
-                        TextButton(onClick = { showServerSignInInfo = true }) {
-                            Text("Why?")
-                        }
-                    }
-                }
-            }
-        }
-    }
-    if (showServerSignInInfo) {
-        AlertDialog(
-            onDismissRequest = { showServerSignInInfo = false },
-            confirmButton = {
-                TextButton(onClick = { showServerSignInInfo = false }) { Text("Got it") }
-            },
-            title = { Text("About server sign-in") },
-            text = {
-                Text(
-                    "BookOrbit controls the available server sign-in options. " +
-                        "Open server sign-in uses the server's own sign-in page and whichever " +
-                        "local or OIDC methods it is configured with. It runs in an embedded " +
-                        "browser inside the app, and some identity providers may block sign-in " +
-                        "from an embedded browser."
-                )
-            }
-        )
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
 private fun LibraryBrowserScreen(
     state: BrowserState,
     onRefresh: () -> Unit,
@@ -1224,398 +949,6 @@ private fun AudioReader(
     }
 }
 
-private data class AudioPlayerSnapshot(
-    val playWhenReady: Boolean,
-    val positionMs: Long,
-    val durationMs: Long?,
-    val speed: Float
-)
-
-private fun audioPlayerSnapshot(session: ReadiumAudioPlaybackService.Session): AudioPlayerSnapshot =
-    AudioPlayerSnapshot(
-        playWhenReady = session.player.playWhenReady,
-        positionMs = session.absolutePositionMs(),
-        durationMs = session.totalDurationMs().takeIf { it > 0L },
-        speed = session.player.playbackParameters.speed
-    )
-
-internal fun shouldShowAudiobookPreviewBanner(launchMode: ReaderLaunchMode): Boolean =
-    launchMode == ReaderLaunchMode.PREVIEW
-
-internal const val BOOK_PREVIEW_MODE_LABEL = "Preview mode · Tap to enable reading progress"
-internal const val AUDIOBOOK_PREVIEW_MODE_LABEL = "Preview mode · Tap to enable listening progress"
-
-@Composable
-internal fun ReadiumCompactAudioPlayer(
-    controller: ReadiumAudioPlaybackController,
-    onClosed: (BookSummary, ReaderLaunchMode) -> Unit = { _, _ -> },
-    onCoverClick: (BookSummary) -> Unit = {},
-    modifier: Modifier = Modifier
-) {
-    val session by produceState<ReadiumAudioPlaybackService.Session?>(null, controller) {
-        controller.session().collect { value = it }
-    }
-    val preparingSession by controller.preparingSession.collectAsState()
-    val currentSession = session
-    if (currentSession == null) {
-        val preparing = preparingSession ?: return
-        val scope = rememberCoroutineScope()
-        Surface(
-            modifier = modifier
-                .fillMaxWidth()
-                .padding(horizontal = 8.dp, vertical = 4.dp)
-                .semantics {
-                    contentDescription = "Preparing audiobook player for ${preparing.book.title}"
-                },
-            shape = RoundedCornerShape(12.dp),
-            color = MaterialTheme.colorScheme.surfaceContainerHigh,
-            tonalElevation = 8.dp,
-            shadowElevation = 8.dp
-        ) {
-            Row(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(8.dp),
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                CompactAudiobookCover(
-                    book = preparing.book,
-                    coverLoader = controller::loadCover
-                )
-                Column(
-                    modifier = Modifier
-                        .weight(1f)
-                        .padding(start = 8.dp)
-                ) {
-                    Text(
-                        preparing.book.title,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                        style = MaterialTheme.typography.titleSmall,
-                        fontWeight = FontWeight.SemiBold
-                    )
-                    Text(
-                        "Preparing audiobook…",
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        style = MaterialTheme.typography.bodySmall
-                    )
-                }
-                IconButton(
-                    onClick = {},
-                    enabled = false,
-                    modifier = Modifier.size(40.dp)
-                ) {
-                    CircularProgressIndicator(
-                        modifier = Modifier.size(22.dp),
-                        strokeWidth = 2.dp,
-                        color = MaterialTheme.colorScheme.primary
-                    )
-                }
-                IconButton(
-                    onClick = {
-                        scope.launch {
-                            controller.close()
-                            onClosed(preparing.book, preparing.launchMode)
-                        }
-                    },
-                    modifier = Modifier.size(36.dp)
-                ) {
-                    Icon(Icons.Default.Close, contentDescription = "Close audiobook player")
-                }
-            }
-        }
-        return
-    }
-    val current = requireNotNull(currentSession)
-    val playback by produceState(audioPlayerSnapshot(current), current.player) {
-        val listener = object : Player.Listener {
-            override fun onEvents(player: Player, events: Player.Events) {
-                value = audioPlayerSnapshot(current)
-            }
-        }
-        current.player.addListener(listener)
-        try {
-            while (isActive) {
-                value = audioPlayerSnapshot(current)
-                delay(500L)
-            }
-        } finally {
-            current.player.removeListener(listener)
-        }
-    }
-    val positionMs = playback.positionMs
-    val durationMs = playback.durationMs
-    val scope = rememberCoroutineScope()
-    var chapterMenuExpanded by remember(current.book.id, current.book.fileId) { mutableStateOf(false) }
-    var speedMenuExpanded by remember(current.book.id, current.book.fileId) { mutableStateOf(false) }
-    var isSeeking by remember(current.book.id, current.book.fileId) { mutableStateOf(false) }
-    var seekPositionMs by remember(current.book.id, current.book.fileId) {
-        mutableStateOf(positionMs.toFloat())
-    }
-    LaunchedEffect(positionMs, isSeeking) {
-        if (!isSeeking) seekPositionMs = positionMs.toFloat()
-    }
-    val chapters = current.book.audioChapters
-    val activeChapterIndex = chapters.indexOfLast { it.startMs <= positionMs }.coerceAtLeast(0)
-    val remainingMs = durationMs
-        ?.minus(if (isSeeking) seekPositionMs.toLong() else positionMs)
-        ?.coerceAtLeast(0L)
-
-    Column(modifier = modifier.fillMaxWidth()) {
-        if (shouldShowAudiobookPreviewBanner(current.launchMode)) {
-            ViewOnlyModeBanner(
-                onConfirmReadMode = controller::enableReadingProgress,
-                modifier = Modifier.padding(horizontal = 8.dp),
-                shape = RoundedCornerShape(12.dp),
-                onPrepareConfirmation = controller::pause,
-                bannerText = AUDIOBOOK_PREVIEW_MODE_LABEL,
-                contentDescription = "Preview mode. Tap to enable listening progress.",
-                dialogText = "You’ll switch to listening mode and start saving your listening progress from this position.",
-                confirmText = "Switch to listening mode"
-            )
-        }
-        Surface(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 8.dp, vertical = 4.dp)
-                .semantics {
-                    contentDescription = "Audiobook player for ${current.book.title}"
-                    stateDescription = if (playback.playWhenReady) "Playing" else "Paused"
-                },
-        shape = RoundedCornerShape(12.dp),
-        color = MaterialTheme.colorScheme.surfaceContainerHigh,
-        tonalElevation = 8.dp,
-        shadowElevation = 8.dp
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            CompactAudiobookCover(
-                book = current.book,
-                coverLoader = controller::loadCover,
-                onClick = { onCoverClick(current.book) }
-            )
-            Column(
-                modifier = Modifier
-                    .weight(1f)
-                    .padding(start = 8.dp)
-            ) {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Column(modifier = Modifier.weight(1f)) {
-                        Text(
-                            current.book.title,
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            style = MaterialTheme.typography.titleSmall,
-                            fontWeight = FontWeight.SemiBold
-                        )
-                        Text(
-                            current.book.author.orEmpty(),
-                            maxLines = 1,
-                            overflow = TextOverflow.Ellipsis,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                            style = MaterialTheme.typography.bodySmall
-                        )
-                    }
-                    IconButton(
-                        onClick = {
-                            scope.launch {
-                                controller.close()
-                                onClosed(current.book, current.launchMode)
-                            }
-                        },
-                        modifier = Modifier.size(36.dp)
-                    ) {
-                        Icon(Icons.Default.Close, contentDescription = "Close audiobook player")
-                    }
-                }
-
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Text(
-                        formatPlaybackTime(if (isSeeking) seekPositionMs.toLong() else positionMs),
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                    Slider(
-                        value = seekPositionMs.coerceIn(0f, (durationMs ?: 1L).toFloat()),
-                        onValueChange = {
-                            isSeeking = true
-                            seekPositionMs = it
-                        },
-                        onValueChangeFinished = {
-                            current.seekToAbsolutePosition(seekPositionMs.toLong())
-                            isSeeking = false
-                        },
-                        valueRange = 0f..(durationMs ?: 1L).toFloat(),
-                        enabled = durationMs != null,
-                        modifier = Modifier
-                            .weight(1f)
-                            .height(32.dp)
-                            .padding(horizontal = 4.dp)
-                    )
-                    Text(
-                        remainingMs?.let { "−${formatPlaybackTime(it)}" } ?: "−–:––",
-                        style = MaterialTheme.typography.labelSmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant
-                    )
-                }
-
-                Row(
-                    modifier = Modifier.fillMaxWidth(),
-                    horizontalArrangement = Arrangement.SpaceEvenly,
-                    verticalAlignment = Alignment.CenterVertically
-                ) {
-                    Box {
-                        IconButton(
-                            onClick = { chapterMenuExpanded = true },
-                            enabled = chapters.isNotEmpty(),
-                            modifier = Modifier.size(40.dp)
-                        ) {
-                            Icon(
-                                Icons.AutoMirrored.Filled.List,
-                                contentDescription = if (chapters.isEmpty()) {
-                                    "No audiobook chapters available"
-                                } else {
-                                    "Select audiobook chapter"
-                                }
-                            )
-                        }
-                        DropdownMenu(
-                            expanded = chapterMenuExpanded,
-                            onDismissRequest = { chapterMenuExpanded = false }
-                        ) {
-                            chapters.forEachIndexed { index, chapter ->
-                                DropdownMenuItem(
-                                    text = {
-                                        Text(
-                                            chapter.title.ifBlank { "Chapter ${index + 1}" },
-                                            fontWeight = if (index == activeChapterIndex) {
-                                                FontWeight.Bold
-                                            } else {
-                                                FontWeight.Normal
-                                            }
-                                        )
-                                    },
-                                    onClick = {
-                                        chapterMenuExpanded = false
-                                        current.seekToAbsolutePosition(chapter.startMs)
-                                    }
-                                )
-                            }
-                        }
-                    }
-                    IconButton(
-                        onClick = {
-                            current.seekToAbsolutePosition(
-                                (current.absolutePositionMs() - 10_000L).coerceAtLeast(0L)
-                            )
-                        },
-                        modifier = Modifier.size(40.dp)
-                    ) {
-                        Icon(Icons.Default.Replay10, contentDescription = "Skip back 10 seconds")
-                    }
-                    IconButton(
-                        onClick = {
-                            if (playback.playWhenReady) current.player.pause() else current.player.play()
-                        },
-                        modifier = Modifier.size(40.dp)
-                    ) {
-                        Icon(
-                            imageVector = if (playback.playWhenReady) Icons.Default.Pause else Icons.Default.PlayArrow,
-                            contentDescription = if (playback.playWhenReady) "Pause audiobook" else "Play audiobook"
-                        )
-                    }
-                    IconButton(
-                        onClick = {
-                            val target = current.absolutePositionMs() + 30_000L
-                            current.seekToAbsolutePosition(
-                                playback.durationMs?.let(target::coerceAtMost) ?: target
-                            )
-                        },
-                        modifier = Modifier.size(40.dp)
-                    ) {
-                        Icon(Icons.Default.Forward30, contentDescription = "Skip forward 30 seconds")
-                    }
-                    Box {
-                        TextButton(
-                            onClick = { speedMenuExpanded = true },
-                            modifier = Modifier
-                                .height(40.dp)
-                                .semantics { contentDescription = "Select playback speed" },
-                            contentPadding = PaddingValues(horizontal = 6.dp)
-                        ) {
-                            Text("${formatPlaybackSpeed(playback.speed.toDouble())}×")
-                        }
-                        DropdownMenu(
-                            expanded = speedMenuExpanded,
-                            onDismissRequest = { speedMenuExpanded = false }
-                        ) {
-                            listOf(0.75, 1.0, 1.25, 1.5, 2.0).forEach { speed ->
-                                DropdownMenuItem(
-                                    text = { Text("${formatPlaybackSpeed(speed)}×") },
-                                    onClick = {
-                                        speedMenuExpanded = false
-                                        controller.setPlaybackSpeed(current.player, speed.toFloat())
-                                    }
-                                )
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-    }
-
-private fun formatPlaybackSpeed(speed: Double) =
-    if (speed % 1.0 == 0.0) speed.toInt().toString() else speed.toString()
-
-@Composable
-private fun CompactAudiobookCover(
-    book: BookSummary,
-    coverLoader: suspend (BookSummary) -> ByteArray?,
-    onClick: () -> Unit = {}
-) {
-    val bitmap by produceState<Bitmap?>(null, book.id, book.coverUrl, book.updatedAtMillis) {
-        value = runCatching { coverLoader(book) }
-            .getOrNull()
-            ?.takeIf { it.isNotEmpty() }
-            ?.let { BitmapFactory.decodeByteArray(it, 0, it.size) }
-    }
-    Box(
-        modifier = Modifier
-            .height(72.dp)
-            .aspectRatio(book.coverAspectRatio.widthToHeight)
-            .background(
-                color = MaterialTheme.colorScheme.primaryContainer,
-                shape = RoundedCornerShape(8.dp)
-            )
-            .clickable(onClick = onClick)
-            .semantics { contentDescription = "Open details for ${book.title}" },
-        contentAlignment = Alignment.Center
-    ) {
-        if (bitmap != null) {
-            Image(
-                bitmap = requireNotNull(bitmap).asImageBitmap(),
-                contentDescription = "Cover for ${book.title}",
-                modifier = Modifier.fillMaxSize(),
-                contentScale = ContentScale.Crop
-            )
-        } else {
-            Text(
-                book.title.take(1).uppercase(),
-                color = MaterialTheme.colorScheme.onPrimaryContainer,
-                style = MaterialTheme.typography.titleLarge
-            )
-        }
-    }
-}
-
 @Composable
 internal fun EpubReaderProgressFooter(
     status: EpubReaderProgressStatus,
@@ -1861,6 +1194,8 @@ private fun ReaderTapZoneTutorialRegion(
         )
     }
 }
+
+internal const val BOOK_PREVIEW_MODE_LABEL = "Preview mode · Tap to enable reading progress"
 
 @Composable
 internal fun ViewOnlyModeBanner(
@@ -2780,18 +2115,6 @@ internal fun percentToChapterIndex(percent: Float?, chapterCount: Int): Int {
     val raw = percent ?: return 0
     val normalized = raw / 100f
     return (normalized.coerceIn(0f, 1f) * (chapterCount - 1)).roundToInt()
-}
-
-private fun formatPlaybackTime(millis: Long): String {
-    val totalSeconds = (millis / 1000L).coerceAtLeast(0L)
-    val hours = totalSeconds / 3600L
-    val minutes = (totalSeconds % 3600L) / 60L
-    val seconds = totalSeconds % 60L
-    return if (hours > 0L) {
-        String.format(Locale.US, "%d:%02d:%02d", hours, minutes, seconds)
-    } else {
-        String.format(Locale.US, "%d:%02d", minutes, seconds)
-    }
 }
 
 internal fun styleEpubHtml(
