+++ app/src/main/java/com/vangeaux/lagrange/BookOrbitRepository.kt
@@ -180,121 +180,6 @@ private enum class AnnotationMutationPostResult {
     DEFERRED
 }
 
-internal fun bookReadingSessionsPath(bookId: String, page: Int, pageSize: Int): String {
-    val encodedBookId = java.net.URLEncoder.encode(bookId, Charsets.UTF_8.name())
-    return "/api/v1/books/$encodedBookId/sessions" +
-        "?page=${page.coerceAtLeast(1)}&pageSize=${pageSize.coerceAtLeast(1)}"
-}
-
-internal fun bookReadingAttemptsPath(bookId: String, page: Int, pageSize: Int): String {
-    val encodedBookId = java.net.URLEncoder.encode(bookId, Charsets.UTF_8.name())
-    return "/api/v1/books/$encodedBookId/reading-attempts" +
-        "?page=${page.coerceAtLeast(1)}&pageSize=${pageSize.coerceAtLeast(1)}"
-}
-
-internal fun readingSessionPath(fileId: String): String {
-    val encodedFileId = java.net.URLEncoder.encode(fileId, Charsets.UTF_8.name())
-    return "/api/v1/books/files/$encodedFileId/sessions"
-}
-
-internal const val ANNOTATIONS_PAGE_SIZE = 30
-
-internal fun annotationsPath(filter: AnnotationsFilter, page: Int, pageSize: Int): String {
-    fun encode(value: String) = java.net.URLEncoder.encode(value, Charsets.UTF_8.name())
-    return buildString {
-        append("/api/v1/annotations?page=${page.coerceAtLeast(1)}&pageSize=${pageSize.coerceAtLeast(1)}")
-        filter.search?.trim()?.takeIf { it.isNotBlank() }?.let { append("&search=").append(encode(it)) }
-        filter.bookId?.trim()?.takeIf { it.isNotBlank() }?.let { append("&bookId=").append(encode(it)) }
-        filter.chapter?.trim()?.takeIf { it.isNotBlank() }?.let { append("&chapter=").append(encode(it)) }
-        filter.colors.takeIf { it.isNotEmpty() }?.let { append("&colors=").append(encode(it.joinToString(","))) }
-        filter.styles.takeIf { it.isNotEmpty() }?.let { append("&styles=").append(encode(it.joinToString(","))) }
-        filter.origins.takeIf { it.isNotEmpty() }?.let { append("&origins=").append(encode(it.joinToString(","))) }
-        filter.dateFrom?.trim()?.takeIf { it.isNotBlank() }?.let { append("&dateFrom=").append(encode(it)) }
-        filter.dateTo?.trim()?.takeIf { it.isNotBlank() }?.let { append("&dateTo=").append(encode(it)) }
-        filter.hasNote?.let { append("&hasNote=").append(it) }
-        append("&status=").append(encode(filter.status))
-        append("&sortBy=").append(encode(filter.sortBy))
-        append("&sortDir=").append(encode(filter.sortDir))
-    }
-}
-
-internal fun annotationsPath(bookId: String): String =
-    "/api/v1/books/${java.net.URLEncoder.encode(bookId, Charsets.UTF_8.name()).replace("+", "%20")}/annotations"
-
-internal fun annotationPath(bookId: String, annotationId: String): String =
-    "${annotationsPath(bookId)}/${java.net.URLEncoder.encode(annotationId, Charsets.UTF_8.name()).replace("+", "%20")}"
-
-internal fun annotationRestorePath(annotationId: String): String {
-    val encodedAnnotationId = java.net.URLEncoder.encode(annotationId, Charsets.UTF_8.name()).replace("+", "%20")
-    return "/api/v1/annotations/$encodedAnnotationId/restore"
-}
-
-internal fun annotationPurgePath(annotationId: String): String {
-    val encodedAnnotationId = java.net.URLEncoder.encode(annotationId, Charsets.UTF_8.name()).replace("+", "%20")
-    return "/api/v1/annotations/$encodedAnnotationId"
-}
-
-internal fun buildCreateAnnotationPayload(item: AnnotationMutationPayload): JSONObject = JSONObject().apply {
-    item.cfi?.let { put("cfi", it) }
-    item.bookFileId?.toIntOrNull()?.let { put("bookFileId", it) }
-    put("text", item.text.orEmpty())
-    put("color", item.color.orEmpty())
-    put("style", item.style.orEmpty())
-    put("note", item.note ?: JSONObject.NULL)
-    put("chapterTitle", item.chapterTitle ?: JSONObject.NULL)
-}
-
-internal fun buildUpdateAnnotationPayload(item: AnnotationMutationPayload): JSONObject = JSONObject().apply {
-    if (item.noteSpecified) put("note", item.note ?: JSONObject.NULL)
-    item.color?.let { put("color", it) }
-    item.style?.let { put("style", it) }
-}
-
-internal fun overlayPendingAnnotations(
-    page: BookAnnotationsPage,
-    pending: List<AnnotationMutationPayload>,
-    mappings: Map<String, String>
-): BookAnnotationsPage {
-    val items = page.items.associateBy { it.id }.toMutableMap()
-    pending.sortedBy { it.updatedAtMillis }.forEach { mutation ->
-        val id = mappings[mutation.annotationId] ?: mutation.annotationId
-        when (mutation.op) {
-            AnnotationMutationOp.CREATE -> if (items[id] == null) {
-                items[id] = BookAnnotation(
-                    id = mutation.annotationId,
-                    bookId = mutation.bookId,
-                    cfi = mutation.cfi,
-                    jumpFileId = mutation.bookFileId,
-                    text = mutation.text,
-                    color = mutation.color,
-                    style = mutation.style,
-                    note = mutation.note,
-                    chapterTitle = mutation.chapterTitle
-                )
-            }
-            AnnotationMutationOp.UPDATE -> items[id]?.let { current ->
-                items[id] = current.copy(
-                    note = if (mutation.noteSpecified) mutation.note else current.note,
-                    color = mutation.color ?: current.color,
-                    style = mutation.style ?: current.style
-                )
-            }
-            AnnotationMutationOp.DELETE -> items.remove(id)
-        }
-    }
-    return page.copy(items = items.values.toList())
-}
-
-internal fun buildReadingSessionPayload(item: ReadingSessionPayload): JSONObject = JSONObject()
-    .put("sessionId", item.sessionId)
-    .put("startedAt", Instant.ofEpochMilli(item.startedAtMillis).toString())
-    .put("endedAt", Instant.ofEpochMilli(item.endedAtMillis).toString())
-    .put("durationSeconds", item.durationSeconds)
-    .apply {
-        item.progressDelta?.let { put("progressDelta", it) }
-        item.endProgress?.let { put("endProgress", it) }
-    }
-
 internal fun extractAccessToken(payload: String): String? {
     val root = runCatching { JSONObject(payload) }.getOrNull() ?: return null
